+++ tests/const_test_suite/const_without_type_test.cpp
@@ -20,8 +20,8 @@ TEST(ConstTest, ConstWithoutTypeFloat)
     interpreter.evaluate(&ast);
 
     ASSERT_TRUE(interpreter.environment->contains("x"));
-    ASSERT_TRUE(is_type<float>(interpreter.environment->get("x")));
-    ASSERT_EQ(as_type<float>(interpreter.environment->get("x")), 4.0);
+    ASSERT_TRUE(is_type<double>(interpreter.environment->get("x")));
+    ASSERT_EQ(as_type<double>(interpreter.environment->get("x")), 4.0);
 }
 
 // INTS
